Optimización del código

+++ src/infrastructure/datasources/resourse.datasourse.impl.ts
@@ -1,13 +1,12 @@
 import { ResourseModel, resourseSchema } from "../../data/mongodb/models/resourse.model";
 import { ResourseDatasource } from "../../domain/datasources";
 import { AddResourseDto, DeleteResourseDto } from "../../domain/dtos/permissions";
-import { DeleteResourseEntity, GetResourseEntity, ResourseEntity } from "../../domain/entities";
+import { DeleteResourseEntity, GetResourseEntity, AddResourseEntity } from "../../domain/entities";
 import { CustomError } from "../../domain/errors/custom.error";
-import { ResourseMapper } from "../mappers";
 
 export class ResourseDatasourceImpl implements ResourseDatasource {
 
-     async addResourse(addResourseDto: AddResourseDto): Promise<ResourseEntity> {
+     async addResourse(addResourseDto: AddResourseDto): Promise<AddResourseEntity> {
           const { resourseName } = addResourseDto;
           try {
                // 1. Crear el resourse
@@ -15,7 +14,7 @@ export class ResourseDatasourceImpl implements ResourseDatasource {
                await resourse.save();
 
                // 2. Mapear la respuesta a la entidad
-               return ResourseMapper.resourseEntityFromObject(resourse);
+               return new AddResourseEntity(resourse.id, resourseName);
 
           } catch (error) {
                throw error;
@@ -24,20 +23,15 @@ export class ResourseDatasourceImpl implements ResourseDatasource {
 
      async deleteResourse(deleteResourseDto: DeleteResourseDto): Promise<DeleteResourseEntity> {
           const { resourseName } = deleteResourseDto;
-          let message: string = "No hay resourse que coincida con la consulta. Se eliminaron 0 resourses."
           try {
+               const exists = await ResourseModel.findOne({ resourseName: resourseName })
+               if (!exists) throw CustomError.badRequest("The resourse name does not exist or has been deleted.")
+
                // 2. eliminar el resourse
                const result = await ResourseModel.deleteOne({ resourseName: resourseName })
 
-               if (result.deletedCount === 1) {
-                    message = "Resourse deleted successfully.";
-               }
-               else {
-                    throw message
-               }
-
                // 3. Mapear la respuesta a la entidad
-               return ResourseMapper.deleteResourseEntityFromObject({ resourseName, message });
+               return new DeleteResourseEntity(resourseName);
 
           } catch (error) {
                throw error;
